Add db root user

+++ srcs/requirements/nginx/tools/init.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
 
-cp /default.conf /etc/nginx/conf.d/default.conf
-sleep 5;
-
+if [ ! -f "/etc/nginx/conf.d/default.conf" ]; then
+	cp /default.conf /etc/nginx/conf.d/default.conf
+	sleep 5;
+fi
 mkdir -p /var/run/nginx
 
 nginx -g 'daemon off;'
